backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers==4.3.1
beautifulsoup4==4.12.3
djangorestframework-simplejwt==5.3.1